+++ numalogic/connectors/rds/__init__.py
@@ -1,0 +1,3 @@
+from numalogic.connectors.rds._rds import RDSFetcher
+
+__all__ = ["RDSFetcher"]
